Added a cache object to store and lookup for diary names when the dia…

+++ src/MySelfLog.Adapter/Mappings/LogBase.cs
@@ -6,7 +6,7 @@ namespace MySelfLog.Adapter.Mappings
     public class LogBase : ICommandFactory
     {
         public int Value { get; protected set; }
-        public int MmolValue { get; protected set; }
+        public decimal MmolValue { get; protected set; }
         public int SlowTerapy { get; protected set; }
         public int FastTerapy { get; protected set; }
         public int Calories { get; protected set; }
@@ -19,7 +19,7 @@ namespace MySelfLog.Adapter.Mappings
             
         }
 
-        public LogBase(int value, int mmolValue, int slowTerapy, int fastTerapy, int calories, string comment,
+        public LogBase(int value, decimal mmolValue, int slowTerapy, int fastTerapy, int calories, string comment,
             IDictionary<string, string> metadata)
         {
             Value = value;
